requirements.txt
fire
numpy
pandas
pillow
plotly
PyYAML
scikit-learn
scipy
seaborn
tabulate
torch
torchvision
tqdm
